Delete unnecessary cookiecutter files
This commit deletes all files from the CookieCutter
Data Science template. If later in the project, it is shown that
any of these files, they can simply be retrieved from the
Cookiecutter Data Science website. Of course, they can also be
retrieved from a previous commit, but note that some of the files
that are added with user-input (i.e., my, Daniel Smarda's input)
were filled with haste, and should probably be regenerated anyway
as the aim and requirements of the project have shifted significantly
since when these files were created.