Add utils file to data directory
This commit is necessary to import relative functions.
Some functions in the future may be repetitive, such as
open_pickle.

Previously, when this code was executed in notebooks, the import
structure was `from ..models import open_pickle`, but that
resulted in an error:

ImportError: attempted relative import with no known parent package.

This commit fixes that error.