Add function to delete results dictionary entries
With the current way file saving is set up,
results with a key that already exist overwrite the value
in the results dictionary, and results with a key that
do not exist are appended to the dictionary. This means that
saving the same value with different names--e.g.
pct_95, pct95, pct_95th--results in different entries.

This function allows developers to delete the excessive
entries.